finishing fetch for volunteers and renaming files
