+++ frontend/src/Components/Admin/Floaters/FloaterToVolunteer.js
@@ -1,4 +1,5 @@
 import React from "react";
+import Message from "../../Message/Message";
 
 export default class VolunteerForm extends React.Component {
     state = { floaters: [], message: false };
@@ -12,18 +13,38 @@ export default class VolunteerForm extends React.Component {
     }
 
     _addFloaterToEvent = floater_id => {
-        console.log(floater_id + " " + this.props.event_id);
+        fetch("/events-floaters", {
+            method: "POST",
+            headers: {
+                Accept: "application/json",
+                "Content-Type": "application/json"
+            },
+            body: JSON.stringify({
+                event_id: this.props.event_id,
+                floater_id: floater_id
+            })
+        })
+            .then(res => res.json())
+            .then(() => {
+                this.setState({ message: true });
+            })
+            .catch(error => console.error(error));
     };
 
     render() {
         return (
             <div className="container mt-2">
+                <Message
+                    show={this.state.message}
+                    status="success"
+                    message="Floater Has Been Added"
+                />
                 <ul className="container list-group mt-4 mb-4">
                     {this.state.floaters.map(floater => {
                         return (
                             <button
                                 type="button"
-                                class="list-group-item list-group-item-action"
+                                className="list-group-item list-group-item-action"
                                 onClick={() => {
                                     this._addFloaterToEvent(floater.floater_id);
                                 }}
@@ -31,7 +52,6 @@ export default class VolunteerForm extends React.Component {
                                 {floater.floater_fname}{" "}
                                 {floater.floater_surname}
                             </button>
-                            // <li className="list-group-item list-group-item-action">{floater.floater_fname}{" "}{floater.floater_surname}</li>
                         );
                     })}
                 </ul>
